Refactor video service and interface, enhance upload flow
Updated `IVideoService` interface and its namespace for better organization. Enhanced `UploadVideoCreate` to include user context and queue messaging. Replaced `DownloadVideoZipRequest` with a simpler `Guid` parameter in download methods for streamlined functionality.
