+++ src/Postech.Fiap.Hackathon.VideoProcessing.WebApi/Features/Videos/Command/UploadVideoCreate.cs
@@ -1,6 +1,6 @@
 using Postech.Fiap.Hackathon.VideoProcessing.WebApi.Common.ResultPattern;
 using Postech.Fiap.Hackathon.VideoProcessing.WebApi.Features.Videos.Contracts;
-using Postech.Fiap.Hackathon.VideoProcessing.WebApi.Features.Videos.Services;
+using Postech.Fiap.Hackathon.VideoProcessing.WebApi.Features.Videos.Interfaces;
 
 namespace Namespace.Postech.Fiap.Hackathon.VideoProcessing.WebApi.Features.Videos.Command;
