+++ src/Postech.Fiap.Hackathon.VideoProcessing.WebApi/Features/Videos/Interfaces/IVideoService.cs
@@ -1,8 +1,7 @@
 using Postech.Fiap.Hackathon.VideoProcessing.WebApi.Common.ResultPattern;
 using Postech.Fiap.Hackathon.VideoProcessing.WebApi.Features.Videos.Contracts;
 
-namespace Postech.Fiap.Hackathon.VideoProcessing.WebApi.Features.Videos.Services;
-
+namespace Postech.Fiap.Hackathon.VideoProcessing.WebApi.Features.Videos.Interfaces;
 
 public interface IVideoService
 {
@@ -10,6 +9,5 @@ public interface IVideoService
 
     Task<Result<UploadVideoResponse>> upload(UploadVideoRequest request, CancellationToken cancellationToken);
 
-    Task<Result<DownloadVideoZipResponse>> download( DownloadVideoZipRequest request, CancellationToken cancellationToken);
+    Task<Result<DownloadVideoZipResponse>> download(Guid id, CancellationToken cancellationToken);
 }
-
